Upload post_topup_cc.sh script
Optional script/sanity check to determine how beneficial performing topup is (i.e., were the susceptibility induced distortions significant). 
The script measures whether b0 after topup is more correlated with T2-weighted scan than the b0's (both in the Neg and Pos direction) before topup. 
This is not an objective measure of topup's performance. 
This script requires the outputdir, T2-weighted scan and brain mask as inputs. 
As an output, it creates a text file (correl.txt) with correlations of b0 with T2-weighted scan before and after topup.