Upload eddy_postproc.sh script
This script does some housekeeping after run_eddy_rodent.sh and then runs dtifit. 
Inputs are the outputdir, scriptdir, and a parameter which specifies if all volumes should be combined (=1) or split (=2). 
If combining volumes, eddy_combine.sh will be needed. This should be contained in whatever you define as your scriptdir.